switch from using socket.io to using Meteor
