+++ .gitignore
@@ -36,7 +36,6 @@ jspm_packages
 # Optional REPL history
 .node_repl_history
 
-# Compiled front-end assets
-client/public/turtlesim-demo.min.css
-client/public/turtlesim-demo.min.js
-client/public/turtlesim-demo.min.js.map
+# Meteor stuff
+.meteor
+settings.json
