--- client/public/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="description" content="A simple little demo to show a simulated turtle drawing shapes"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Turtlesim Shape Demo</title>
-    <link rel="stylesheet" href="turtlesim-demo.min.css" />
-    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.0.2/dist/leaflet.css" />
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Raleway:100,400,600" />
-  </head>
-  <body>
-    <h1>Turtlesim Shape Demo</h1>
-    <p>A simple little demo to show a simulated turtle drawing shapes</p>
-    <div class="draw-stars">
-      <button class="star-button" id="draw-star-1-turtle">Draw a star (1 turtle)</button>
-      <button class="star-button" id="draw-star-2-turtles">Draw a star (2 turtles)</button>
-    </div>
-    <p>Choose an SVG file to see it in action.</p>
-    <div>
-      <input type="file" id="fileinput" name="fileinput" class="fileinput" />
-      <label for="fileinput">Choose a file</label>
-    </div>
-    <div id="turtlemap" class="map-container"></div>
-    <script src="turtlesim-demo.min.js"></script>
-  </body>
-</html>
