--- client/scss/_variables.scss
@@ -1,4 +1,0 @@
-$white: #fff;
-$dark-gray: #333;
-$blue: #006bb6;
-$dark-blue: #004a80;
