--- client/scss/turtlesim-demo.scss
@@ -1,2 +1,0 @@
-@import 'variables';
-@import 'base';
