+++ client/star-paths.json
@@ -1,0 +1,34 @@
+{
+  "oneTurtleStar": [
+    [
+      [4, 0],
+      [3, 3],
+      [0, 3],
+      [2.5, 5],
+      [1.5, 8],
+      [4, 6],
+      [6.5, 8],
+      [5.5, 5],
+      [8, 3],
+      [5, 3],
+      [4, 0]
+    ]
+  ],
+  "twoTurtlesStar": [
+    [
+      [4, 0],
+      [3, 3],
+      [0, 3],
+      [2.5, 5],
+      [1.5, 8],
+      [4, 6]
+    ], [
+      [4, 6],
+      [6.5, 8],
+      [5.5, 5],
+      [8, 3],
+      [5, 3],
+      [4, 0]
+    ]
+  ]
+}
