+++ client/svg-tools.js
